Додані об'єкти, які потрібні для вироблення енергетики: 1. Сировина (уран, торій-239) 2. Продукція (уран-238, уран-235, плутоній-239, торієвий стрижень, урановий стрижень) 3. Уранова шахта, у планах торієву додати 4. Збагачувальний завод 5. Уранова АЕС 6. Розрахунки для енергетики
Змінено:
1. Заводи, вихід продукції зроблено у словнику
2. розрахунки для заводів

+++ Tycoonia/Application/Factory/LaunchControleCenterFactory.cs
@@ -33,7 +33,7 @@ namespace Tycoonia.Application.Factory
             }
         }
 
-        public static Dictionary<string, int> CreateBufferCheck(FactoryBase factory, StorageResources storageResources, PlayerReal player,  int expectedOutput)
+        public static Dictionary<string, long> CreateBufferCheck(FactoryBase factory, StorageResources storageResources, PlayerReal player,  int expectedOutput)
         {
             foreach (var item in factory.ReceipeList)
             {
